🚚 rename:	ChapterPage 스타일 코드 분리 (#13)
